small correction on ROS_WARN %lu not valid
correction for fake moveit with niryo one

small corrections on launch files in niryo_robot_bringup

correction on urdf for niryo one incorrect

+++ niryo_robot_hardware_stack/can_driver/src/can_interface_core.cpp
@@ -546,7 +546,7 @@ void CanInterfaceCore::addSingleCommandToQueue(const std::shared_ptr<common::mod
             // keep position cmd apart
             if (_conveyor_cmds.size() > QUEUE_OVERFLOW)
             {
-                ROS_WARN("CanInterfaceCore::addCommandToQueue: Cmd queue overflow ! %lu", _conveyor_cmds.size());
+                ROS_WARN("CanInterfaceCore::addCommandToQueue: Cmd queue overflow ! %d", static_cast<int>(_conveyor_cmds.size()));
             }
             else
             {
@@ -557,7 +557,7 @@ void CanInterfaceCore::addSingleCommandToQueue(const std::shared_ptr<common::mod
         {
             if (_stepper_single_cmds.size() > QUEUE_OVERFLOW)
             {
-                ROS_WARN("CanInterfaceCore::addCommandToQueue: Cmd queue overflow ! %lu", _stepper_single_cmds.size());
+                ROS_WARN("CanInterfaceCore::addCommandToQueue: Cmd queue overflow ! %d", static_cast<int>(_stepper_single_cmds.size()));
             }
             else
             {
