+++ niryo_robot_hardware_stack/ttl_driver/src/ttl_interface_core.cpp
@@ -876,7 +876,7 @@ void TtlInterfaceCore::addSingleCommandToQueue(const std::shared_ptr<common::mod
     if (cmd->isValid())
     {
         if (_single_cmds_queue.size() > QUEUE_OVERFLOW)
-            ROS_WARN("TtlInterfaceCore::addSingleCommandToQueue: dxl cmd queue overflow ! %lu", _single_cmds_queue.size());
+            ROS_WARN("TtlInterfaceCore::addSingleCommandToQueue: dxl cmd queue overflow ! %d", static_cast<int>(_single_cmds_queue.size()));
 
         if (cmd->isDxlCmd())
             _single_cmds_queue.push(std::dynamic_pointer_cast<common::model::DxlSingleCmd>(cmd));
@@ -885,7 +885,7 @@ void TtlInterfaceCore::addSingleCommandToQueue(const std::shared_ptr<common::mod
             if (cmd->getCmdType() == static_cast<int>(EStepperCommandType::CMD_TYPE_CONVEYOR))
             {
                 if (_conveyor_cmds_queue.size() > QUEUE_OVERFLOW)
-                    ROS_WARN("TtlInterfaceCore::addCommandToQueue: Cmd queue overflow ! %lu", _conveyor_cmds_queue.size());
+                    ROS_WARN("TtlInterfaceCore::addCommandToQueue: Cmd queue overflow ! %d", static_cast<int>(_conveyor_cmds_queue.size()));
                 else
                     _conveyor_cmds_queue.push(std::dynamic_pointer_cast<common::model::StepperTtlSingleCmd>(cmd));
             }
